pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "jobu"
version = "0.1.0"
description = "Python 기반 통합 배치 스케줄링 시스템"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "dok9"}
]
keywords = ["batch", "scheduler", "cron", "job", "async"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: AsyncIO",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "aiosqlite>=0.20.0,<0.22.0",
    "aiosql>=13.0,<14.0",
    "pyyaml>=6.0.0,<7.0",
    "pydantic>=2.10.0,<3.0",
    "croniter>=5.0.0,<7.0",
    "fastapi>=0.115.0,<0.130.0",
    "uvicorn>=0.34.0,<0.40.0",
    "jinja2>=3.1.0,<4.0",
    "httpx>=0.27.0,<0.30.0",
    "asyncpg>=0.30.0,<0.32.0",
    "asyncmy>=0.2.9,<0.3.0",
    "python-json-logger>=3.0.0,<5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<10.0",
    "pytest-asyncio>=1.0.0,<2.0",
]

[project.urls]
Homepage = "https://github.com/dok9gold/jobu"
Repository = "https://github.com/dok9gold/jobu"
Issues = "https://github.com/dok9gold/jobu/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["admin*", "common*", "database*", "dispatcher*", "worker*"]

[tool.setuptools.package-data]
"*" = ["*.sql", "*.yaml.example", "*.html"]
